backend/requirements.txt

# NOTE: ChromaDB depends on native wheels (e.g. chroma-hnswlib) that are not
# currently published for Python 3.14 on Windows.
# Use Python 3.11–3.13 for a smooth install.

fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-dotenv>=1.2.0
python-multipart>=0.0.9
langchain-ollama>=1.0.1
langgraph>=1.0.7
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
pydantic>=2.7.0
# Pin to avoid resolver backtracking into chromadb<=0.5.x (which forces numpy<2
# and triggers a source build on newer Python versions).
chromadb==0.6.3
starlette>=0.38.0